requirements.txt
absl-py==0.15.0
altgraph==0.17.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
cycler==0.11.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageai==2.1.6
importlib-metadata==4.12.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
pefile==2022.5.30
Pillow==7.0.0
protobuf==3.9.2
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
PySimpleGUI==4.60.3
python-dateutil==2.8.2
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
six==1.15.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.8.1
